Add traffic sign classification and handling; refactor Tsr and ActuatorKuksa for improved traffic sign processing

+++ apps/rpi5-firmware/actuator/ActuatorKuksa.cpp
@@ -27,10 +27,10 @@ void ActuatorKuksa::setCruiseControl(const bool flag, const int targetSpeed) {
 
 void ActuatorKuksa::setTrafficSign(const int trafficSign) {
 	ActuatorDecorator::setTrafficSign(trafficSign);
-	_kuksa.sendValueToKuksa("Vehicle.ADAS.TrafficSignRecognition.DetectedSignType", (int)trafficSign);
+	_kuksa.sendValueToKuksa("Vehicle.ADAS.TrafficSignRecognition.DetectedSignType", static_cast<uint8_t>(trafficSign));
 }
 
 void ActuatorKuksa::setSpeedLimit(const int speedLimit) {
 	ActuatorDecorator::setSpeedLimit(speedLimit);
-	_kuksa.sendValueToKuksa("Vehicle.ADAS.TrafficSignRecognition.DetectedSpeedLimit", (int)speedLimit);
+	_kuksa.sendValueToKuksa("Vehicle.ADAS.TrafficSignRecognition.DetectedSpeedLimit", static_cast<float>(speedLimit));
 }
